[Enhance] refactor todo-item - deleteHandler 제거 (이제 prop으로 받음)
